+++ app/Resources/views/base.html.twig
@@ -7,17 +7,9 @@
         <link rel="icon" type="image/x-icon" href="{{ asset('favicon.ico') }}" />
     </head>
     <body>
-        ______________________________________________
-        <br />
-        header<br />
-        ______________________________________________
-        <br/>
+       
         {% block body %}{% endblock %}
-        ___________________________________________
-        <br />
-        footer<br />
-        __________________________________________
-        <br />
+       
         {% block javascripts %}{% endblock %}
     </body>
 </html>
