+++ app/config/config.yml
@@ -7,11 +7,11 @@ imports:
 # Put parameters here that don't need to change on each machine where the app is deployed
 # http://symfony.com/doc/current/best_practices/configuration.html#application-related-configuration
 parameters:
-    locale: en
+    locale: fr
 
 framework:
     #esi:             ~
-    #translator:      { fallbacks: ["%locale%"] }
+    translator:      { fallbacks: ["%locale%"] }
     secret:          "%secret%"
     router:
         resource: "%kernel.root_dir%/config/routing.yml"
@@ -75,3 +75,25 @@ assetic:
         cssrewrite: ~
         sass: ~
         compass: ~
+
+
+#Easy Admin
+easy_admin:
+    design:
+        #color_scheme: 'light'
+       # menu: ['User', 'Page']
+    entities:
+        User:
+            class: HliveBundle\Entity\User           
+        "Pages":
+            class: HliveBundle\Entity\Page
+            form:
+                fields:
+                    - titre
+                    - slug
+                    - { property: description, type: ckeditor }
+#FosUserBudle
+fos_user:
+    db_driver: orm 
+    firewall_name: main
+    user_class: HliveBundle\Entity\User
