+++ app/config/routing.yml
@@ -5,3 +5,11 @@ hlive:
 app:
     resource: "@AppBundle/Controller/"
     type:     annotation
+
+easy_admin_bundle:
+    resource: "@EasyAdminBundle/Controller/"
+    type:     annotation
+    prefix:   /admin
+
+fos_user:
+    resource: "@FOSUserBundle/Resources/config/routing/all.xml"
