+++ nbproject/private/private.properties
@@ -1,4 +1,6 @@
+copy.src.files=false
 copy.src.on.open=false
+copy.src.target=
 index.file=index.php
 run.as=LOCAL
-url=http://localhost/hlive/Symfony
+url=http://localhost/hlive/
