+++ nbproject/private/private.xml
@@ -3,14 +3,12 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/var/www/hlive/Symfony/src/HliveBundle/Resources/config/doctrine/Annonce.orm.yml</file>
-            <file>file:/var/www/hlive/Symfony/src/HliveBundle/Entity/Annonce.php</file>
-            <file>file:/var/www/hlive/Symfony/src/HliveBundle/Entity/Bien.php</file>
-            <file>file:/var/www/hlive/Symfony/src/HliveBundle/Entity/Categorie.php</file>
-            <file>file:/var/www/hlive/Symfony/src/HliveBundle/Entity/Ville.php</file>
-            <file>file:/var/www/hlive/Symfony/src/HliveBundle/Resources/config/doctrine/Ville.orm.yml</file>
-            <file>file:/var/www/hlive/Symfony/src/HliveBundle/Resources/config/doctrine/Categorie.orm.yml</file>
-            <file>file:/var/www/hlive/Symfony/src/HliveBundle/Resources/config/doctrine/Bien.orm.yml</file>
+            <file>file:/C:/wamp64/www/hlive/src/HliveBundle/Resources/config/doctrine/Page.orm.yml</file>
+            <file>file:/C:/wamp64/www/hlive/src/HliveBundle/Entity/User.php</file>
+            <file>file:/C:/wamp64/www/hlive/src/HliveBundle/Resources/config/services.yml</file>
+            <file>file:/C:/wamp64/www/hlive/src/HliveBundle/Resources/config/routing.yml</file>
+            <file>file:/C:/wamp64/www/hlive/app/config/routing.yml</file>
+            <file>file:/C:/wamp64/www/hlive/src/HliveBundle/Resources/config/doctrine/User.orm.yml</file>
         </group>
     </open-files>
 </project-private>
