--- src/HliveBundle/Entity/Annonce.php
@@ -1,199 +1,0 @@
-<?php
-
-namespace HliveBundle\Entity;
-
-/**
- * Annonce
- */
-class Annonce
-{
-    /**
-     * @var integer
-     */
-    private $id;
-
-    /**
-     * @var string
-     */
-    private $titre;
-
-    /**
-     * @var string
-     */
-    private $contenu;
-
-    /**
-     * @var \DateTime
-     */
-    private $date_add;
-
-    /**
-     * @var \DateTime
-     */
-    private $date_update;
-
-    /**
-     * @var boolean
-     */
-    private $active;
-
-    /**
-     * @var \HliveBundle\Entity\Bien
-     */
-    private $bien;
-
-
-    /**
-     * Get id
-     *
-     * @return integer
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    /**
-     * Set titre
-     *
-     * @param string $titre
-     *
-     * @return Annonce
-     */
-    public function setTitre($titre)
-    {
-        $this->titre = $titre;
-
-        return $this;
-    }
-
-    /**
-     * Get titre
-     *
-     * @return string
-     */
-    public function getTitre()
-    {
-        return $this->titre;
-    }
-
-    /**
-     * Set contenu
-     *
-     * @param string $contenu
-     *
-     * @return Annonce
-     */
-    public function setContenu($contenu)
-    {
-        $this->contenu = $contenu;
-
-        return $this;
-    }
-
-    /**
-     * Get contenu
-     *
-     * @return string
-     */
-    public function getContenu()
-    {
-        return $this->contenu;
-    }
-
-    /**
-     * Set dateAdd
-     *
-     * @param \DateTime $dateAdd
-     *
-     * @return Annonce
-     */
-    public function setDateAdd($dateAdd)
-    {
-        $this->date_add = $dateAdd;
-
-        return $this;
-    }
-
-    /**
-     * Get dateAdd
-     *
-     * @return \DateTime
-     */
-    public function getDateAdd()
-    {
-        return $this->date_add;
-    }
-
-    /**
-     * Set dateUpdate
-     *
-     * @param \DateTime $dateUpdate
-     *
-     * @return Annonce
-     */
-    public function setDateUpdate($dateUpdate)
-    {
-        $this->date_update = $dateUpdate;
-
-        return $this;
-    }
-
-    /**
-     * Get dateUpdate
-     *
-     * @return \DateTime
-     */
-    public function getDateUpdate()
-    {
-        return $this->date_update;
-    }
-
-    /**
-     * Set active
-     *
-     * @param boolean $active
-     *
-     * @return Annonce
-     */
-    public function setActive($active)
-    {
-        $this->active = $active;
-
-        return $this;
-    }
-
-    /**
-     * Get active
-     *
-     * @return boolean
-     */
-    public function getActive()
-    {
-        return $this->active;
-    }
-
-    /**
-     * Set bien
-     *
-     * @param \HliveBundle\Entity\Bien $bien
-     *
-     * @return Annonce
-     */
-    public function setBien(\HliveBundle\Entity\Bien $bien = null)
-    {
-        $this->bien = $bien;
-
-        return $this;
-    }
-
-    /**
-     * Get bien
-     *
-     * @return \HliveBundle\Entity\Bien
-     */
-    public function getBien()
-    {
-        return $this->bien;
-    }
-}
