+++ src/HliveBundle/Entity/User.php~
@@ -1,0 +1,14 @@
+<?php
+
+namespace HliveBundle\Entity;
+
+use FOS\UserBundle\Model\User as BaseUser;
+use Doctrine\ORM\Mapping as ORM;
+
+/**
+ * User
+ */
+class User extends BaseUser
+{
+    
+}
