--- src/HliveBundle/Entity/Ville.php
@@ -1,71 +1,0 @@
-<?php
-
-namespace HliveBundle\Entity;
-
-/**
- * Ville
- */
-class Ville
-{
-    /**
-     * @var integer
-     */
-    private $id;
-
-
-    /**
-     * Get id
-     *
-     * @return integer
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-    /**
-     * @var \Doctrine\Common\Collections\Collection
-     */
-    private $biens;
-
-    /**
-     * Constructor
-     */
-    public function __construct()
-    {
-        $this->biens = new \Doctrine\Common\Collections\ArrayCollection();
-    }
-
-    /**
-     * Add bien
-     *
-     * @param \HliveBundle\Entity\Bien $bien
-     *
-     * @return Ville
-     */
-    public function addBien(\HliveBundle\Entity\Bien $bien)
-    {
-        $this->biens[] = $bien;
-
-        return $this;
-    }
-
-    /**
-     * Remove bien
-     *
-     * @param \HliveBundle\Entity\Bien $bien
-     */
-    public function removeBien(\HliveBundle\Entity\Bien $bien)
-    {
-        $this->biens->removeElement($bien);
-    }
-
-    /**
-     * Get biens
-     *
-     * @return \Doctrine\Common\Collections\Collection
-     */
-    public function getBiens()
-    {
-        return $this->biens;
-    }
-}
