--- src/HliveBundle/Resources/config/doctrine/Categorie.orm.yml
@@ -1,35 +1,0 @@
-HliveBundle\Entity\Categorie:
-    type: entity
-    table: categorie
-    #repositoryClass: HliveBundle\Repository\CommentaireRepository
-    id:
-        id:
-            type: integer
-            id: true
-            generator:
-                strategy: AUTO
-    fields:
-        nom:
-            type: string
-            nullable: false
-            length: 255
-            fixed: false
-            comment: ''
-        description:
-            type: text
-            nullable: true
-            fixed: false
-            comment: ''
-        active:
-            type: boolean
-            nullable: false
-            length: 5
-            fixed: false
-            option:
-               default: 0
-    manyToOne:
-        categorieParent:
-           targetEntity: Categorie
-           inversedBy: enfantCategories
-        
-    lifecycleCallbacks: {  }
