+++ src/HliveBundle/Resources/config/doctrine/User.orm.yml
@@ -1,0 +1,9 @@
+HliveBundle\Entity\User:
+    type:  entity
+    table: user
+    id:
+        id:
+            type: integer
+            generator:
+                strategy: AUTO                
+    lifecycleCallbacks: {  }
