--- src/HliveBundle/Resources/config/doctrine/Ville.orm.yml
@@ -1,14 +1,0 @@
-HliveBundle\Entity\Localisation:
-    type: entity
-    table: localisation
-    #repositoryClass: HliveBundle\Repository\CommentaireRepository
-    id:
-        id:
-            type: integer
-            id: true
-            generator:
-                strategy: AUTO
-    fields:
-       
-        
-    lifecycleCallbacks: {  }
