+++ src/HliveBundle/Resources/config/routing.yml
@@ -1,3 +1,22 @@
 hlive_homepage:
     path: /hello/{nom}/{prenom}
     defaults: { _controller: HliveBundle:Default:index }
+    
+#fos_user_security:
+#    resource: routing/security.yml
+#    prefix: /
+#fos_user_profile:
+#    resource: routing/profil.yml
+#    prefix: /compte
+
+#fos_user_register:
+#    resource: routing/inscription.yml
+#    prefix: /inscription
+
+#fos_user_resetting:
+#    resource: routing/resetting.yml
+#    prefix: /resetting
+
+#fos_user_change_password:
+#    resource: routing/change_mot_passe.yml
+#    prefix: /compte
