+++ src/HliveBundle/Resources/config/routing/change_mot_passe.yml
@@ -1,0 +1,3 @@
+fos_user_change_password:
+    pattern: /changer-mot-de-passe
+    defaults: { _controller: FOSUserBundle:ChangePassword:changePassword }
