+++ src/HliveBundle/Resources/config/routing/groupe.yml
@@ -1,0 +1,20 @@
+fos_user_group_list:
+    pattern: /liste
+    defaults: { _controller: FOSUserBundle:Group:list }
+
+fos_user_group_new:
+    pattern: /new
+    defaults: { _controller: FOSUserBundle:Group:new }
+
+fos_user_group_show:
+    pattern: /{groupName}
+    defaults: { _controller: FOSUserBundle:Group:show }
+
+fos_user_group_edit:
+    pattern: /{groupName}/edit
+    defaults: { _controller: FOSUserBundle:Group:edit }
+
+fos_user_group_delete:
+    pattern: /{groupName}/delete
+    defaults: { _controller: FOSUserBundle:Group:delete }
+
