+++ src/HliveBundle/Resources/config/routing/inscription.yml
@@ -1,0 +1,3 @@
+fos_user_registration_register:
+    pattern: /
+    defaults: { _controller: FOSUserBundle:Registration:register }
